+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Implement MROUND(n, m)
 * Implement MROUND2(n, m)
 * Allow for fun names starting with _ (for lambdas mostly)
+* Implement CEILING(n, m)
 
 
 ## xel.js 1.3.0  released 2024-04-05